Split what grew past the limits the structure gate keeps
CI's structure gate refused the branch on three counts and every one was mine.
Local runs of the crate's own tests said nothing about it, because the gate is
a separate script and I had not run it.

`supervise` grew a select block and passed the function limit. The restart
work's race against cancellation is a unit of its own — it answers "was a stop
requested while we were putting this back" and returns that as a bool — so it
moves out as `restore_for_restart`, and the supervision loop and its helper
move together into `session/supervise.rs`. A session's lifecycle and the loop
that watches a running one are different lengths of thing to read.

The session tests passed the file limit, so they split by what they ask:
what becomes of a running session's children, and what a session settles
before it has any.

The contract-digest test passed the function limit because most of its body
had become a table of one edit per covered field. That is data, and a test
body that is mostly data stops reading as a statement about behaviour, so the
table is named and lifted beside it.

Nothing here changes behaviour. Confirmed by running the gate itself rather
than inferring it: check-structure and its self-test pass, along with the six
other guards CI runs, workspace clippy is clean under `-D warnings`, and 2016
tests pass.
